requirements.txt
streamlit>=1.32
supabase>=2.3
pandas>=2.1
plotly>=5.18
python-dotenv>=1.0
